Reorganize into src and resource module folders
