+++ .gitignore
@@ -1,3 +1,3 @@
 /.idea/
 /out/
-/classes/*.class
+/classes/
