--- classes/images/checkersRules.html
@@ -1,79 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<HTML>
-<HEAD>
-	<META HTTP-EQUIV="CONTENT-TYPE" CONTENT="text/html; charset=utf-8">
-	<TITLE></TITLE>
-
-	<META NAME="AUTHOR" CONTENT="Jason Hamilton">
-	<STYLE TYPE="text/css">
-	<!--
-		@page { margin: 0.2in }
-		P { margin-bottom: 0.2in }
-	-->
-	</STYLE>
-</HEAD>
-
-<BODY LANG="en-US" DIR="LTR">
-<P ALIGN=CENTER STYLE="margin-bottom: 0in"><FONT SIZE=5>Checkers
-Rules (U.S.)</FONT></P>
-<P STYLE="margin-bottom: 0in"><BR>
-</P>
-<OL>
-	<LI><P STYLE="margin-bottom: 0in">Checkers is played by two players.
-	Each player begins the game with 12 colored discs. (Typically, one
-	set of pieces is black and the other red.)</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">The board consists of 64 squares,
-	alternating between 32 dark and 32 light squares. It is positioned
-	so that each player has a light square on the right side corner
-	closest to him or her.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">Each player places his or her
-	pieces on the 12 dark squares closest to him or her.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">Black moves first. Players then
-	alternate moves.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">Moves are allowed only on the dark
-	squares, so pieces always move diagonally. Single pieces are always
-	limited to forward moves (toward the opponent).</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">A piece making a non-capturing
-	move (not involving a jump) may move only one square.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">A piece making a capturing move (a
-	jump) leaps over one of the opponent's pieces, landing in a straight
-	diagonal line on the other side. Only one piece may be captured in a
-	single jump; however, multiple jumps are allowed on a single turn.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">When a piece is captured, it is
-	removed from the board.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">If a player is able to make a
-	capture, there is no option -- the jump must be made. If more than
-	one capture is available, the player is free to choose whichever he
-	or she prefers.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">When a piece reaches the furthest
-	row from the player who controls that piece, it is crowned and
-	becomes a king. One of the pieces which had been captured is placed
-	on top of the king so that it is twice as high as a single piece.</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">Kings are limited to moving
-	diagonally, but may move both forward and backward. (Remember that
-	single pieces, i.e. non-kings, are always limited to forward moves.)</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P STYLE="margin-bottom: 0in">Kings may combine jumps in several
-	directions -- forward and backward -- on the same turn. Single
-	pieces may shift direction diagonally during a multiple capture
-	turn, but must always jump forward (toward the opponent).</P>
-	<P STYLE="margin-bottom: 0in"></P>
-	<LI><P>A player wins the game when the opponent cannot make a move.
-	In most cases, this is because all of the opponent's pieces have
-	been captured, but it could also be because all of his pieces are
-	blocked in.</P>
-</OL>
-<P STYLE="margin-bottom: 0in"><BR>
-</P>
-</BODY>
-</HTML>
