+++ resources/black.xcf
