+++ resources/board.xcf
