+++ resources/red.xcf
